Feat: completed strategic npc, moved Player class from poker_game.py to player.py
