+++ src/poker_game.py
@@ -1,119 +1,6 @@
-import random
 from dataclasses import dataclass
-from deck import Deck, Card
-from rank import BestHand
-from typing import Union
-
-
-@dataclass
-class Player:
-    name: str = random.sample(["Henry", "Jiachen", "Jonathan"], 1)[0]
-    stack: int = 0
-    hand = []
-    bet_this_round: int = 0
-    current_decision: Union[int, str, None] = None
-    status: Union[
-        str, None
-    ] = None  # 'little blind', 'big blind', 'dealer' 'highest bettor'
-
-    def __repr__(self) -> str:
-        return str((self.name, self.stack, self.hand))
-
-    def draw_hand(self, deck: Deck) -> None:
-        self.hand = [deck.draw() for _ in range(2)]
-
-    def clear_hand(self) -> None:
-        self.hand = []
-
-    def best_hand(self, board: list[Card]) -> BestHand:
-        return BestHand(self.hand, board)
-
-    def reset_action(self):
-        self.status = None
-        self.current_decision = None
-        self.bet_this_round = 0
-
-    def bet(self, action: int) -> None:
-        """
-        Sets current_decision and bet_this_round attributs on self
-        """
-        self.current_decision = action
-        self.bet_this_round += self.current_decision
-
-    def decision(self, current_bet: int, pot: int, board: list[Card]) -> None:
-        """
-        Decision method prompts the user to choose an action for their turn.
-        """
-        # Skip player who is all-in
-        if self.stack == 0:
-            self.current_decision = 0
-            return
-
-        possible_actions = {"ALL IN", "FOLD"}
-        if current_bet > self.bet_this_round:
-            possible_actions.add("CALL")
-            if self.stack >= current_bet * 2:
-                action = input(
-                    f"The board is now {*board,}."
-                    f"{self.name}'s hand is {self.hand}, with ${self.stack} chips. "
-                    f"The current pot is ${pot}. The current bet is ${current_bet}. "
-                    f"You have put in ${self.bet_this_round} already. "
-                    f"Please make a decision between: {*possible_actions,}"
-                    f", or enter your raise size from ${current_bet * 2 - self.bet_this_round} to ${self.stack} "
-                )
-            else:
-                action = input(
-                    f"The board is now {*board,}."
-                    f"{self.name}'s hand is {self.hand}, with ${self.stack} chips. "
-                    f"The current pot is ${pot}. The current bet is ${current_bet}. "
-                    f"You have put in ${self.bet_this_round} already. "
-                    f"Please make a decision between: {*possible_actions,}."
-                )
-        elif current_bet == self.bet_this_round:
-            possible_actions.add("CHECK")
-            action = input(
-                f"The board is now {*board,}."
-                f"{self.name}'s hand is {self.hand}, with ${self.stack} chips. "
-                f"The current pot is ${pot}. The current bet is ${current_bet}. "
-                f"You have put in ${self.bet_this_round} already. "
-                f"Please make a decision in: {*possible_actions,}"
-                f", or enter your bet size from ${current_bet + 1} to ${self.stack} "
-            )
-
-        try:
-            action = int(action)
-            if current_bet * 2 <= action <= self.stack or action == current_bet:
-                self.bet(action - self.bet_this_round)
-            else:
-                print("Please bet a valid amount")
-                self.decision(current_bet, pot)
-        except ValueError:
-            if action.upper() == "ALL IN":
-                self.bet(self.stack)
-            elif action.upper() == "FOLD":
-                self.current_decision = "FOLD"
-            elif action.upper() == "CHECK":
-                if current_bet == self.bet_this_round:
-                    self.current_decision = 0
-                    self.status = None
-                else:
-                    print("You cannot check")
-                    self.decision(current_bet, pot)
-            elif action.upper() == "CALL":
-                if self.bet_this_round < current_bet <= self.stack:
-                    self.bet(current_bet - self.bet_this_round)
-                elif current_bet == self.bet_this_round:
-                    print("You cannot call")
-                    self.decision(current_bet, pot)
-                elif current_bet > self.stack:
-                    print("Your stack is less than the current bet, doing ALL IN instead.")
-                    self.bet(self.stack)
-                # for testing, to be removed
-                else:
-                    assert False, "something is wrong"
-            else:
-                print("Invalid decision.")
-                self.decision(current_bet, pot)
+from deck import Deck
+from player import Player
 
 
 @dataclass
